requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.0
pypdf>=3.17.0
python-docx>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
numpy>=1.24.0

